окончательно определился со скоростями. На лёгкой: 15 кл/с для игрока, 16 кл/с для врага. На средней в 1.4 раза больше, на сложной в 1.8, на молнии в 2.2. Пересчитал моменты спавна частиц в соответствии с новыми скоростями, хотя расписания для сложностей со средней и выше нуждаются в доработке

+++ My_Nedo_Pakman/Game_logic.cpp
@@ -46,7 +46,7 @@ void Enemy::init(int number, int difficulty) {
 	y_ = start_enemy_coords[number][Y_coord];
 	lastX = x_; lastY = y_;
 	speed = enemy_speed[difficulty];
-	millisec_in_cell = 1000. / (double)speed;
+	millisec_in_cell = 1000. / speed;
 }
 
 void Enemy::start(clock_t current_time) {
